fix float error

+++ app/api/gotoposition.py
@@ -65,7 +65,7 @@ def go_to_position_():
 
     path = pathfinding_application_service.find(
         obj_obstacles, width, lenght, robot_position, destination_position,
-        ROBOT_RADIUS / SCALING)
+        int(ROBOT_RADIUS / SCALING))
     path = get_segments.get_filter_path(path)
     upscale_path = []
     print("PATHHHHHHHHHHHHHHHHHHHHHHHH")
